fix Again ratings with FSRS
Stability now decreases after Again rating (previously it stayed the same). This
was a bug when implementing the algorithm.

As a result removed the forced "Hard" rating after an "Again" rating, and
removed the notion of "isFsrsIntroduced" since the way mistake tracking is
implemented no longer created "Again" ratings for skills that have never been
introduced.