use posthog-js for web
On Web posthog-react-native wasn't sending `$pageview` and it wasn't capturing
properties like `$current_url`, `$host`, `$pathname`, `$search_engine`,
`$referrer`, `$referring_domain`. So instead of maintaining a patches on
posthog-react-native I'll just use posthog-js for web.

The downside of this is that on Web page views are captured as `$pageview`
events but on native they're captured as `$screen` events.